fix UnboundLocalError when no source is sent
git-svn-id: https://repos.nextthought.com/svn/nti-svn/nti.app.products.courseware/trunk@94915 3dfb626b-079f-4ab1-a9f6-4654e5377508

+++ src/nti/app/products/courseware/importer.py
@@ -61,6 +61,7 @@ def _execute(course, archive_path, writeout=True):
 	if course is None:
 		raise ValueError("Invalid course")
 
+	tmp_path = None
 	try:
 		tmp_path = check_archive(archive_path)
 		filer = DirectoryFiler(tmp_path or archive_path)
@@ -68,7 +69,8 @@ def _execute(course, archive_path, writeout=True):
 		result = importer.process(course, filer, writeout)
 		return result
 	finally:
-		delete_dir(tmp_path)
+		if tmp_path:
+			delete_dir(tmp_path)
 
 def import_course(ntiid, archive_path, writeout=True):
 	"""
